+++ scripts/jsdom.js
@@ -1,0 +1,9 @@
+/* eslint-env mocha */
+
+beforeEach(function () {
+  this.jsdom = require('jsdom-global')();
+});
+
+afterEach(function () {
+  this.jsdom();
+});
