+++ src/widgets/toggle/toggle.js
@@ -1,18 +1,22 @@
-import find from 'lodash/collection/find';
-import React from 'react';
-import ReactDOM from 'react-dom';
 import {
   bemHelper,
-  prepareTemplateProps,
   getContainerNode
 } from '../../lib/utils.js';
+import defaultTemplates from './defaultTemplates.js';
 import cx from 'classnames';
 import autoHideContainerHOC from '../../decorators/autoHideContainer.js';
 import headerFooterHOC from '../../decorators/headerFooter.js';
-import defaultTemplates from './defaultTemplates.js';
 import RefinementListComponent from '../../components/RefinementList/RefinementList.js';
+import currentToggle from './implementations/currentToggle';
+import legacyToggle from './implementations/legacyToggle';
 
 const bem = bemHelper('ais-toggle');
+
+// we cannot use helper. because the facet is not yet declared in the helper
+const hasFacetsRefinementsFor = (attributeName, searchParameters) =>
+  searchParameters &&
+  searchParameters.facetsRefinements &&
+  searchParameters.facetsRefinements[attributeName] !== undefined;
 
 /**
  * Instantiate the toggling of a boolean facet filter on and off.
@@ -76,18 +80,18 @@ function toggle({
   } = {}) {
   const containerNode = getContainerNode(container);
 
+  if (!container || !attributeName || !label) {
+    throw new Error(usage);
+  }
+
   let RefinementList = headerFooterHOC(RefinementListComponent);
   if (autoHideContainer === true) {
     RefinementList = autoHideContainerHOC(RefinementList);
   }
 
-  if (!container || !attributeName || !label) {
-    throw new Error(usage);
-  }
-
   const hasAnOffValue = userValues.off !== undefined;
 
-  let cssClasses = {
+  const cssClasses = {
     root: cx(bem(null), userCssClasses.root),
     header: cx(bem('header'), userCssClasses.header),
     body: cx(bem('body'), userCssClasses.body),
@@ -100,104 +104,37 @@ function toggle({
     count: cx(bem('count'), userCssClasses.count)
   };
 
+  // store the computed options for usage in the two toggle implementations
+  const implemOptions = {
+    attributeName,
+    label,
+    userValues,
+    templates,
+    collapsible,
+    transformData,
+    hasAnOffValue,
+    containerNode,
+    RefinementList,
+    cssClasses
+  };
+
   return {
-    getConfiguration: () => ({
-      disjunctiveFacets: [attributeName]
-    }),
-    init({state, helper, templatesConfig}) {
-      this._templateProps = prepareTemplateProps({
-        transformData,
-        defaultTemplates,
-        templatesConfig,
-        templates
-      });
-      this.toggleRefinement = this.toggleRefinement.bind(this, helper);
+    getConfiguration(currentSearchParameters, searchParametersFromUrl) {
+      const useLegacyToggle =
+        hasFacetsRefinementsFor(attributeName, currentSearchParameters) ||
+        hasFacetsRefinementsFor(attributeName, searchParametersFromUrl);
 
-      // no need to refine anything at init if no custom off values
-      if (!hasAnOffValue) {
-        return;
-      }
-      // Add filtering on the 'off' value if set
-      const isRefined = state.isDisjunctiveFacetRefined(attributeName, userValues.on);
-      if (!isRefined) {
-        helper.addDisjunctiveFacetRefinement(attributeName, userValues.off);
-      }
+      const toggleImplementation = useLegacyToggle ?
+        legacyToggle(implemOptions) :
+        currentToggle(implemOptions);
+
+      this.init = toggleImplementation.init.bind(toggleImplementation);
+      this.render = toggleImplementation.render.bind(toggleImplementation);
+      return toggleImplementation.getConfiguration(currentSearchParameters, searchParametersFromUrl);
     },
-    toggleRefinement: (helper, facetValue, isRefined) => {
-      const on = userValues.on;
-      const off = userValues.off;
-
-      // Checking
-      if (!isRefined) {
-        if (hasAnOffValue) {
-          helper.removeDisjunctiveFacetRefinement(attributeName, off);
-        }
-        helper.addDisjunctiveFacetRefinement(attributeName, on);
-      } else {
-        // Unchecking
-        helper.removeDisjunctiveFacetRefinement(attributeName, on);
-        if (hasAnOffValue) {
-          helper.addDisjunctiveFacetRefinement(attributeName, off);
-        }
-      }
-
-      helper.search();
-    },
-    render({helper, results, state, createURL}) {
-      const isRefined = helper.state.isDisjunctiveFacetRefined(attributeName, userValues.on);
-      const onValue = userValues.on;
-      const offValue = userValues.off === undefined ? false : userValues.off;
-      const allFacetValues = results.getFacetValues(attributeName);
-      const onData = find(allFacetValues, {name: onValue.toString()});
-      const onFacetValue = {
-        name: label,
-        isRefined: onData !== undefined ? onData.isRefined : false,
-        count: onData === undefined ? null : onData.count
-      };
-      const offData = find(allFacetValues, {name: offValue.toString()});
-      const offFacetValue = {
-        name: label,
-        isRefined: offData !== undefined ? offData.isRefined : false,
-        count: offData === undefined ? null : offData.count
-      };
-
-      // what will we show by default,
-      // if checkbox is not checked, show: [ ] free shipping (countWhenChecked)
-      // if checkbox is checked, show: [x] free shipping (countWhenNotChecked)
-      const nextRefinement = isRefined ? offFacetValue : onFacetValue;
-
-      const facetValue = {
-        name: label,
-        isRefined,
-        count: nextRefinement === undefined ? null : nextRefinement.count,
-        onFacetValue,
-        offFacetValue
-      };
-
-      // Bind createURL to this specific attribute
-      function _createURL() {
-        return createURL(
-          state
-            .removeDisjunctiveFacetRefinement(attributeName, isRefined ? onValue : userValues.off)
-            .addDisjunctiveFacetRefinement(attributeName, isRefined ? userValues.off : onValue)
-        );
-      }
-
-      ReactDOM.render(
-        <RefinementList
-          collapsible={collapsible}
-          createURL={_createURL}
-          cssClasses={cssClasses}
-          facetValues={[facetValue]}
-          shouldAutoHideContainer={results.nbHits === 0}
-          templateProps={this._templateProps}
-          toggleRefinement={this.toggleRefinement}
-        />,
-        containerNode
-      );
-    }
+    init() {},
+    render() {}
   };
 }
-
 
 export default toggle;
